requirements.txt
Flask==2.3.2
scikit-learn==1.2.2
nltk==3.8.1
joblib==1.2.0
pandas==2.1.0
gunicorn==20.1.0
python-dotenv==1.0.0
